python/pyproject.toml
[project]
name = "vision-forge"
version = "0.1.0"
description = "Visual neural network builder - local desktop version"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "ForgeOpus", email = "contact@forgeopus.dev"}
]
keywords = ["deep-learning", "neural-networks", "pytorch", "tensorflow", "visual-programming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
# AI features - optional for users who want local AI assistance
ai = [
    "google-generativeai>=0.8.3",
    "anthropic>=0.39.0",
    "pillow>=12.0.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
vision-forge = "vision_forge.cli:main"

[project.urls]
Homepage = "https://github.com/ForgeOpus/visionforge"
Documentation = "https://github.com/ForgeOpus/visionforge/blob/main/README.md"
Repository = "https://github.com/ForgeOpus/visionforge"
Issues = "https://github.com/ForgeOpus/visionforge/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vision_forge"]

[tool.hatch.build.targets.wheel.force-include]
"vision_forge/web" = "vision_forge/web"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
